backend/requirement.txt
fastapi>=0.68.0
uvicorn>=0.15.0
google-cloud-vision>=3.0.0
langchain-groq>=0.1.0
python-multipart>=0.0.5
python-dotenv>=0.19.0